chore(endpoints): move thunks
