+++ ui/src/resources/components/GetResources.tsx
@@ -7,7 +7,7 @@ import {getAuthorizations} from 'src/authorizations/actions/thunks'
 import {getBuckets} from 'src/buckets/actions/thunks'
 import {getChecks} from 'src/checks/actions'
 import {getDashboards} from 'src/dashboards/actions/thunks'
-import {getEndpoints} from 'src/notifications/endpoints/actions'
+import {getEndpoints} from 'src/notifications/endpoints/actions/thunks'
 import {getLabels} from 'src/labels/actions'
 import {getMembers} from 'src/members/actions/thunks'
 import {getNotificationRules} from 'src/notifications/rules/actions'
